pyproject.toml
[project]
name = "guidance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "arize-phoenix>=12.9.0",
    "boto3>=1.34.0",
    "fastapi>=0.120.4",
    "litellm>=1.79.1",
    "llama-index>=0.14.7",
    "openinference-instrumentation-litellm>=0.1.27",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.44",
    "sse-starlette>=2.2.1",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.3",
]
